remove duplication from sqlite schema helpers
