+++ packages/knex/src/dialects/sqlite/index.ts
@@ -1,4 +1,5 @@
 export * from './BaseSqlitePlatform';
+export * from './BaseSqliteSchemaHelper';
 export * from './SqliteTableCompiler';
 export * from './BetterSqliteKnexDialect';
 export * from './LibSqlKnexDialect';
